Prevent cross-origin frame SecurityError when referer different domain.
The GoldenLayoutService uses the presence of window.opener to work out
if it's a child window. This breaks when an app is arrived at as a
redirect from a different domain.

This patch makes golden layout additional check that the opener window
is not cross-origin when working out if it's in a child window.